+++ .github/test-stacks/python/__main__.py
@@ -5,4 +5,6 @@ config = pulumi.Config()
 name = config.require("name")
 random_host_name = random.RandomPet(name)
 
+pulumi.debug("I'm a random pet", random_host_name)
+
 pulumi.export("name", random_host_name.id)