+++ src/config.ts
@@ -100,13 +100,15 @@ export function makeConfig() {
       diff: getBooleanInput('diff'),
       refresh: getBooleanInput('refresh'),
       replace: parseSemicolorToArray(getMultilineInput('replace')),
+      exclude: parseSemicolorToArray(getMultilineInput('exclude')),
+      excludeDependents: getBooleanInput('exclude-dependents'),
       target: parseSemicolorToArray(getMultilineInput('target')),
       targetDependents: getBooleanInput('target-dependents'),
       policyPacks: parseSemicolorToArray(getMultilineInput('policyPacks')),
       policyPackConfigs: parseSemicolorToArray(
         getMultilineInput('policyPackConfigs'),
       ),
-      userAgent: 'pulumi/actions@v5',
+      userAgent: 'pulumi/actions@v6',
       color: getUnionInput('color', {
         alternatives: ['always', 'never', 'raw', 'auto'] as const,
       }),
@@ -115,6 +117,10 @@ export function makeConfig() {
       suppressOutputs: getBooleanInput('suppress-outputs'),
       suppressProgress: getBooleanInput('suppress-progress'),
       continueOnError: getBooleanInput('continue-on-error'),
+      logVerbosity: getNumberInput('log-verbosity'),
+      logFlow: getBooleanInput('log-flow'),
+      logToStdErr: !!getInput('log-verbosity'),  // logToStdErr is true if logVerbosity has a truthy value
+      debug: getBooleanInput('debug'),
     },
   };
 }